// api/api.ts
export interface Cooldown {
    /**
     * The symbol of the ship that is on cooldown
     * @type {string}
     * @memberof Cooldown
     */
    'shipSymbol': string;
    /**
     * The total duration of the cooldown in seconds
     * @type {number}
     * @memberof Cooldown
     */
    'totalSeconds': number;
    /**
     * The remaining duration of the cooldown in seconds
     * @type {number}
     * @memberof Cooldown
     */
    'remainingSeconds': number;
    /**
     * The date and time when the cooldown expires in ISO 8601 format
     * @type {string}
     * @memberof Cooldown
     */
    'expiration'?: string;
}
// ...
export interface Ship {
    /**
     * The globally unique identifier of the ship in the following format: `[AGENT_SYMBOL]-[HEX_ID]`
     * @type {string}
     * @memberof Ship
     */
    'symbol': string;
    /**
     * 
     * @type {ShipRegistration}
     * @memberof Ship
     */
    'registration': ShipRegistration;
    /**
     * 
     * @type {ShipNav}
     * @memberof Ship
     */
    'nav': ShipNav;
    /**
     * 
     * @type {ShipCrew}
     * @memberof Ship
     */
    'crew': ShipCrew;
    /**
     * 
     * @type {ShipFrame}
     * @memberof Ship
     */
    'frame': ShipFrame;
    /**
     * 
     * @type {ShipReactor}
     * @memberof Ship
     */
    'reactor': ShipReactor;
    /**
     * 
     * @type {ShipEngine}
     * @memberof Ship
     */
    'engine': ShipEngine;
    /**
     * 
     * @type {Cooldown}
     * @memberof Ship
     */
    'cooldown': Cooldown;
    /**
     * Modules installed in this ship.
     * @type {Array<ShipModule>}
     * @memberof Ship
     */
    'modules': Array<ShipModule>;
    /**
     * Mounts installed in this ship.
     * @type {Array<ShipMount>}
     * @memberof Ship
     */
    'mounts': Array<ShipMount>;
    /**
     * 
     * @type {ShipCargo}
     * @memberof Ship
     */
    'cargo': ShipCargo;
    /**
     * 
     * @type {ShipFuel}
     * @memberof Ship
     */
    'fuel': ShipFuel;
}
/**
 * Ship cargo details.
 * @export
 * @interface ShipCargo
 */
export interface ShipCargo {
    /**
     * The max number of items that can be stored in the cargo hold.
     * @type {number}
     * @memberof ShipCargo
     */
    'capacity': number;
    /**
     * The number of items currently stored in the cargo hold.
     * @type {number}
     * @memberof ShipCargo
     */
    'units': number;
    /**
     * The items currently in the cargo hold.
     * @type {Array<ShipCargoItem>}
     * @memberof ShipCargo
     */
    'inventory': Array<ShipCargoItem>;
}
/**
 * The type of cargo item and the number of units.
 * @export
 * @interface ShipCargoItem
 */
export interface ShipCargoItem {
    /**
     * 
     * @type {TradeSymbol}
     * @memberof ShipCargoItem
     */
    'symbol': TradeSymbol;
    /**
     * The name of the cargo item type.
     * @type {string}
     * @memberof ShipCargoItem
     */
    'name': string;
    /**
     * The description of the cargo item type.
     * @type {string}
     * @memberof ShipCargoItem
     */
    'description': string;
    /**
     * The number of units of the cargo item.
     * @type {number}
     * @memberof ShipCargoItem
     */
    'units': number;
}


/**
 * The ship\'s crew service and maintain the ship\'s systems and equipment.
 * @export
 * @interface ShipCrew
 */
export interface ShipCrew {
    /**
     * The current number of crew members on the ship.
     * @type {number}
     * @memberof ShipCrew
     */
    'current': number;
    /**
     * The minimum number of crew members required to maintain the ship.
     * @type {number}
     * @memberof ShipCrew
     */
    'required': number;
    /**
     * The maximum number of crew members the ship can support.
     * @type {number}
     * @memberof ShipCrew
     */
    'capacity': number;
    /**
     * The rotation of crew shifts. A stricter shift improves the ship\'s performance. A more relaxed shift improves the crew\'s morale.
     * @type {string}
     * @memberof ShipCrew
     */
    'rotation': ShipCrewRotationEnum;
    /**
     * A rough measure of the crew\'s morale. A higher morale means the crew is happier and more productive. A lower morale means the ship is more prone to accidents.
     * @type {number}
     * @memberof ShipCrew
     */
    'morale': number;
    /**
     * The amount of credits per crew member paid per hour. Wages are paid when a ship docks at a civilized waypoint.
     * @type {number}
     * @memberof ShipCrew
     */
    'wages': number;
}

export const ShipCrewRotationEnum = {
    Strict: 'STRICT',
    Relaxed: 'RELAXED'
} as const;

export type ShipCrewRotationEnum = typeof ShipCrewRotationEnum[keyof typeof ShipCrewRotationEnum];

/**
 * The engine determines how quickly a ship travels between waypoints.
 * @export
 * @interface ShipEngine
 */
export interface ShipEngine {
    /**
     * The symbol of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'symbol': ShipEngineSymbolEnum;
    /**
     * The name of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'name': string;
    /**
     * The description of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'description': string;
    /**
     * Condition is a range of 0 to 100 where 0 is completely worn out and 100 is brand new.
     * @type {number}
     * @memberof ShipEngine
     */
    'condition'?: number;
    /**
     * The speed stat of this engine. The higher the speed, the faster a ship can travel from one point to another. Reduces the time of arrival when navigating the ship.
     * @type {number}
     * @memberof ShipEngine
     */
    'speed': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipEngine
     */
    'requirements': ShipRequirements;
}

export const ShipEngineSymbolEnum = {
    ImpulseDriveI: 'ENGINE_IMPULSE_DRIVE_I',
    IonDriveI: 'ENGINE_ION_DRIVE_I',
    IonDriveIi: 'ENGINE_ION_DRIVE_II',
    HyperDriveI: 'ENGINE_HYPER_DRIVE_I'
} as const;

export type ShipEngineSymbolEnum = typeof ShipEngineSymbolEnum[keyof typeof ShipEngineSymbolEnum];

/**
 * The frame of the ship. The frame determines the number of modules and mounting points of the ship, as well as base fuel capacity. As the condition of the frame takes more wear, the ship will become more sluggish and less maneuverable.
 * @export
 * @interface ShipFrame
 */
export interface ShipFrame {
    /**
     * Symbol of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'symbol': ShipFrameSymbolEnum;
    /**
     * Name of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'name': string;
    /**
     * Description of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'description': string;
    /**
     * Condition is a range of 0 to 100 where 0 is completely worn out and 100 is brand new.
     * @type {number}
     * @memberof ShipFrame
     */
    'condition'?: number;
    /**
     * The amount of slots that can be dedicated to modules installed in the ship. Each installed module take up a number of slots, and once there are no more slots, no new modules can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'moduleSlots': number;
    /**
     * The amount of slots that can be dedicated to mounts installed in the ship. Each installed mount takes up a number of points, and once there are no more points remaining, no new mounts can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'mountingPoints': number;
    /**
     * The maximum amount of fuel that can be stored in this ship. When refueling, the ship will be refueled to this amount.
     * @type {number}
     * @memberof ShipFrame
     */
    'fuelCapacity': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipFrame
     */
    'requirements': ShipRequirements;
}
// ...
export type ShipFrameSymbolEnum = typeof ShipFrameSymbolEnum[keyof typeof ShipFrameSymbolEnum];

/**
 * Details of the ship\'s fuel tanks including how much fuel was consumed during the last transit or action.
 * @export
 * @interface ShipFuel
 */
export interface ShipFuel {
    /**
     * The current amount of fuel in the ship\'s tanks.
     * @type {number}
     * @memberof ShipFuel
     */
    'current': number;
    /**
     * The maximum amount of fuel the ship\'s tanks can hold.
     * @type {number}
     * @memberof ShipFuel
     */
    'capacity': number;
    /**
     * 
     * @type {ShipFuelConsumed}
     * @memberof ShipFuel
     */
    'consumed'?: ShipFuelConsumed;
}
// ...
export interface ShipModule {
    /**
     * The symbol of the module.
     * @type {string}
     * @memberof ShipModule
     */
    'symbol': ShipModuleSymbolEnum;
    /**
     * Modules that provide capacity, such as cargo hold or crew quarters will show this value to denote how much of a bonus the module grants.
     * @type {number}
     * @memberof ShipModule
     */
    'capacity'?: number;
    /**
     * Modules that have a range will such as a sensor array show this value to denote how far can the module reach with its capabilities.
     * @type {number}
     * @memberof ShipModule
     */
    'range'?: number;
    /**
     * Name of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'name': string;
    /**
     * Description of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'description': string;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipModule
     */
    'requirements': ShipRequirements;
}
// ...
export type ShipModuleSymbolEnum = typeof ShipModuleSymbolEnum[keyof typeof ShipModuleSymbolEnum];

/**
 * A mount is installed on the exterier of a ship.
 * @export
 * @interface ShipMount
 */
export interface ShipMount {
    /**
     * Symbo of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'symbol': ShipMountSymbolEnum;
    /**
     * Name of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'name': string;
    /**
     * Description of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'description'?: string;
    /**
     * Mounts that have this value, such as mining lasers, denote how powerful this mount\'s capabilities are.
     * @type {number}
     * @memberof ShipMount
     */
    'strength'?: number;
    /**
     * Mounts that have this value denote what goods can be produced from using the mount.
     * @type {Array<string>}
     * @memberof ShipMount
     */
    'deposits'?: Array<ShipMountDepositsEnum>;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipMount
     */
    'requirements': ShipRequirements;
}
// ...
export type ShipMountSymbolEnum = typeof ShipMountSymbolEnum[keyof typeof ShipMountSymbolEnum];
// ...
export interface ShipNav {
    /**
     * The symbol of the system.
     * @type {string}
     * @memberof ShipNav
     */
    'systemSymbol': string;
    /**
     * The symbol of the waypoint.
     * @type {string}
     * @memberof ShipNav
     */
    'waypointSymbol': string;
    /**
     * 
     * @type {ShipNavRoute}
     * @memberof ShipNav
     */
    'route': ShipNavRoute;
    /**
     * 
     * @type {ShipNavStatus}
     * @memberof ShipNav
     */
    'status': ShipNavStatus;
    /**
     * 
     * @type {ShipNavFlightMode}
     * @memberof ShipNav
     */
    'flightMode': ShipNavFlightMode;
}
// ...
export type ShipNavStatus = typeof ShipNavStatus[keyof typeof ShipNavStatus];


/**
 * The reactor of the ship. The reactor is responsible for powering the ship\'s systems and weapons.
 * @export
 * @interface ShipReactor
 */
export interface ShipReactor {
    /**
     * Symbol of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'symbol': ShipReactorSymbolEnum;
    /**
     * Name of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'name': string;
    /**
     * Description of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'description': string;
    /**
     * Condition is a range of 0 to 100 where 0 is completely worn out and 100 is brand new.
     * @type {number}
     * @memberof ShipReactor
     */
    'condition'?: number;
    /**
     * The amount of power provided by this reactor. The more power a reactor provides to the ship, the lower the cooldown it gets when using a module or mount that taxes the ship\'s power.
     * @type {number}
     * @memberof ShipReactor
     */
    'powerOutput': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipReactor
     */
    'requirements': ShipRequirements;
}
// ...
export interface ShipRegistration {
    /**
     * The agent\'s registered name of the ship
     * @type {string}
     * @memberof ShipRegistration
     */
    'name': string;
    /**
     * The symbol of the faction the ship is registered with
     * @type {string}
     * @memberof ShipRegistration
     */
    'factionSymbol': string;
    /**
     * 
     * @type {ShipRole}
     * @memberof ShipRegistration
     */
    'role': ShipRole;
}